Refactor: Move `_buildFeedItem` and `_buildBody` out of `build` method
The `_buildFeedItem` and `_buildBody` methods in `ProfileView` were previously defined within the `build` method. This commit refactors them to be top-level methods within the class for better organization and readability.

Additionally, a settings icon button has been added to the `AppBarComponent` in the `ProfileView`, which navigates to the settings page when pressed. The `buildBody` method is now wrapped in an `Obx` widget to react to changes in the controller.